Change some stuff
Angular momentum and memory are removed. Trials are now one-on-one.

+++ src/scalar.hpp
@@ -3,7 +3,7 @@
 
 namespace bouncers {
 
-typedef double scalar;
+typedef float scalar;
 
 } /* namespace bouncers */
